Refactor settings panel for Artifact frame
